// REST_API/src/parser/parser.interface.ts
/*
	This is a API_NKAIDH project file.
	Developer: sizeoftrickster <sizeoftrickster@gmail.com>
	
	See more here https://github.com/sizeoftrickster/API_NKAIDH
	
	Copyright (c) 2024 Andrey Gazizov <sizeoftrickster@gmail.com>. All rights reserved.
*/

export interface IParseGroupColumns
{
	groupNumber: number;
	columnStart: string;
	columnEnd: string;
}; // export interface IParseGroupColumns

export interface IParseGroupRows
{
	day: string;
	rowStart: number;
	rowEnd: number;
}; // export interface IParseGroupRows