// REST_API/src/types/parser.d.ts
/*
	This is a API_NKAIDH project file.
	Developer: sizeoftrickster <sizeoftrickster@gmail.com>
	
	See more here https://github.com/sizeoftrickster/API_NKAIDH
	
	Copyright (c) 2024 Andrey Gazizov <sizeoftrickster@gmail.com>. All rights reserved.
*/

export type TDirection = "RC" | "AT";
export type TDayIndex = 0 | 1 | 2 | 3 | 4 | 5 | 6;
export type TDayString = "ПОНЕДЕЛЬНИК" | "ВТОРНИК" | "СРЕДА" | "ЧЕТВЕРГ" | "ПЯТНИЦА" | "СУББОТА";